Changing social to use Chi's instead
