+++ frontend/pages/signup.py
@@ -79,23 +79,19 @@ def register(_, username, email, password):
     elif password == '':
         return html.Div('Password is empty, try again')
 
-    add_response = None
-    try:
-        add_response = util.sendRequest(parameters=["create_user", username, email, password])
-        print(add_response)
-    except:
-        return html.Div('An error occurred while running the useradd script')
+    add_response = util.sendRequest(parameters=["create_user", username, email, password])
     # Return the response in HTML
-    if add_response.get("returnCode") == "1":
+    if add_response.get("returnCode") == 1:
         dcc.Store(id='stored-userid', 
                 data=add_response.get("userid"), 
                 storage_type='session')
-        return dcc.Location(pathname='/logSucc', id='redirect')
-    if add_response.get("returnCode") == "2":
-        return html.Div('Invalid login, try again',
+        return dcc.Location(pathname='/login', id='redirect')
+    elif add_response.get("returnCode") == -1:
+        return html.Div('Username or email is already in use. Please try again.',
                          style={'color': 'red'})
     else:
         return html.Div('Unhandled error', style={'color': 'red'})
+    # return html.Div(add_response.get("returnCode"));
 
 def layout():
     return signuppage
