+++ frontend/pages/social.py
@@ -1,21 +1,22 @@
 import os
 import sys
-import pipes
-import math
-import numpy as np
 
-from datetime import datetime
-
+from stat import FILE_ATTRIBUTE_NO_SCRUB_DATA
 import dash
-from dash import Input, Output, State, callback, dcc, html, no_update
-#from dash_extensions.enrich import MultiplexerTransform 
+from dash import Dash, html, dcc, callback, Input, Output, State, no_update
+import pandas as pd
+import plotly.figure_factory as ff
+import plotly.express as px
+import plotly.graph_objects as go
+import requests as r
+import json
+import datetime as dt
+from datetime import date
 
 #Relative path import for util.py
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 import util
 
-# Didn't want to use this but I just need to get this done now
-# global_chats = []
 
 dash.register_page(
     __name__, 
@@ -23,229 +24,233 @@ dash.register_page(
     path='/social'
 )
 
-# Layout: Success (Temporary)
-success = html.Div(id='main-content', children=[
-    html.Div(id='userid-text', style={'display': 'none'}),
-    html.Button('Create Chat', id='create-chat', className='create-button'),
-    html.Div(id='content-area', children=[
-        html.Div(id='chatroom-table', className='hidden'),
-        html.Div(id='chat-table', className='hidden'),
-        # html.Div(id='error-text'),
-        html.Div(id='friends-table', className='hidden'),
-        html.Div(id='chat-id', style={'display': 'none'})
-    ]),
-    dcc.Interval(
-        id='interval',
-        interval=1*1000, # in milliseconds
-        n_intervals=0
-    )
-])
+friends_list = ['John', 'Frank', 'Eric', 'Bob', 'Dylan', 'Shawn', 'Shane', 'Kobe Bryant', 'Hi', 'Hello', 'Hola', 'Shalom']
+blocked_list = ['Sonjay']
 
-# Pull the userid session
+# Generate people list (friends/blocked)
+def create_people_div_list(people_list):
+    # people_div_list = []
+    # friend_item_style = {
+    #               'height':'10%',
+    #               'width':'100%',
+    #               'borderBottom':'1px solid black',
+    #               'display':'block',
+    #               'alignItems':'center'}
+    # for person in people_list:
+    #     if person in friends_list:
+    #       people_div_list.append(
+    #         html.Div([
+    #           person,
+    #           html.Button(['Chat'], id=f'{person}-chatbtn'),
+    #           html.Button(['Block'], id=f'{person}-blockbtn')
+    #           ], style=friend_item_style),
+    #       )
+    #     else:
+    #       people_div_list.append(
+    #         html.Div([
+    #           person,
+    #           html.Button(id=f'{person}-chatbtn', style={'display':'none'}),
+    #           html.Button(['Unblock'], id=f'{person}-blockbtn')
+    #           ], style=friend_item_style),
+    #       )
+    # return people_div_list
+    util.sendRequest(parameters=["get_user_rooms"])
+    return html.Div("Testing")
+
+# App layout
+layout = html.Div([
+  html.Div([
+    html.Div([
+      html.Div('User Chats', 
+        style={
+          'backgroundColor':'#2f4f04',
+          'height':'10%',
+          'color':'white'
+      }),
+      html.Div(create_people_div_list(util),
+        id='friends-list',
+        style={
+          'overflowY':'scroll',
+          'width':'100%',
+          'height':'90%',
+          'maxHeight':'90%'
+        }
+      )
+    ],
+      id='list-type', 
+      style={
+        'height':'50vh',
+        'maxHeight':'60%',
+        'width':'100%',
+        'border':'1px solid black',
+    }),
+    html.Div([
+      html.Div('Blocked List', 
+        style={
+          'backgroundColor':'#2f4f04',
+          'height':'10%',
+          'color':'white'
+      }),
+      html.Div(create_people_div_list(blocked_list),
+        id='blocked-list',
+        style={
+          'overflowY':'scroll',
+          'width':'100%',
+          'height':'90%',
+          'maxHeight':'90%'
+        }
+      )
+    ], 
+      style={
+        'marginTop':'20px',
+        'height':'30vh',
+        'maxHeight':'30%',
+        'width':'100%',
+        'border':'1px solid black',
+    }),
+  ], style={
+    'margin':'auto',
+    'maxWidth':'90%',
+    'minWidth':'90%',
+    'maxHeight':'90%',
+    'minHeight':'90%',
+  }),
+  html.Div([
+      html.Div([
+        'Choose a recipient'
+      ], id='recipient', style={
+        'height':'10%',
+        'width':'100%',
+        'borderBottom':'1px solid black',
+        'backgroundColor':'#2f4f04',
+        'color':'white',
+        'fontWeight':'bold',
+        'fontSize':'30px',
+      }),
+      html.Div([
+        html.Div([
+          
+        ], id='messages',
+                 style={
+                   'overflow':'auto',
+                   'maxHeight':'80%'
+                 })
+      ], id='conversation',
+          style={
+        'width':'100%',
+        'height':'80%',
+        'borderBottom':'1px solid black',
+      }),
+      html.Div([
+        dcc.Input(
+          id='send_message_field',
+          style={
+          'height':'50%',
+          'width':'80%',
+          'marginLeft': '10px'
+        }),
+        html.Button(['Send'], 
+          id='send_message_btn', 
+          style={
+          'height':'50%',
+          'width':'10%',
+          'marginLeft':'5px'
+        })
+      ], style={
+        'width':'100%',
+        'height':'10%',
+        'display':'flex',
+        'alignItems':'center'
+      })
+    ], style={
+        'margin':'auto',
+        'width':'80%',
+        'height':'90%',
+        'border':'1px solid black',
+    }),
+], style={
+  'display':'grid',
+  'gridTemplateColumns':'1fr 1fr',
+  'width':'95%',
+  'height':'95vh',
+  'margin':'auto'
+})
+
 @dash.callback(
-    Output('userid-text', 'children'),
-    Output('chatroom-table', 'className'),
-    Output('chat-table', 'className'),
-    Output('friends-table', 'className'),
-    Input('create-chat', 'n_clicks'),
-    State('session-userid', 'data'),
-    State('content-area', 'className'),
+    Output('area', 'children'),
+    Output('lat', 'children'),
+    Output('lon', 'children'),
+    Input('map', 'clickData'),
     prevent_initial_call=True
 )
-def setid(_, data, className):
-    print(data)
-    return data, '', '', ''
-
-# Join a chat
+def click(clickdata):
+    print(clickdata)
+    points = clickdata['points'][0]
+    print(points['lat'])
+    print(points['lon'])
+    print(points['customdata'][0])
+    print(points)
+    return points['customdata'][0], points['lat'], points['lon']
+  
 @dash.callback(
-    Output('chat-table', 'children'),
-    Output('create-chat', 'style'),
-    Output('chat-id', 'value'),
-    Input('userid-text', 'children'),
-    prevent_initial_call=True
-)
-def join(userid):
-    response = None
-    # try:
-        # TODO: Implement dynamic area and time elements with the graph area gui
-        # Need to fix input, both inputs will be broken with certain characters and spaces
-    print('pre chat request')
-    response = util.sendRequest("join_group", "Mountains", "9AM", userid)
-    print('post c r')
-    # except:
-        # return html.Div('An error occurred while running the createGroup1 script')
-    r_c = response['returnCode']
+    Output('output-container-date-picker-single', 'children'),
+    Input('my-date-picker-single', 'date'),
+    State('time-dropdown', 'value'),
+    prevent_initial_call=True)
+def update_output(date_value, time):
+    string_prefix = 'You have selected: '
+    if time is not None:
+      if date_value is not None:
+          date_object = date.fromisoformat(date_value)
+          date_string = date_object.strftime('%B %d, %Y')
+          return string_prefix + date_string + ' at ' + time
 
-    # 3 - 5 are failure errors
-    if r_c == 3 or r_c == 4 or r_c == 5:
-        return html.Div(response['message'], style={'color': 'red'}), {'display': 'none'}, '-1'
-    return [
-        html.Div(children=[
-            html.H1(response['area'], id='area-title'),
-            html.P(response['time'], id='area-time')
-        ], id='messages-area'),
-        html.Table(children=[html.Tbody(children=[], id='table-body')], id='messages-table'),
-        html.Div(id='input-area', 
-        children=[
-            
-            dcc.Input('', className='input', id='message-input', debounce=True, type='text'),
-        ])
-    ], {'display': 'none'}, response['chatid']
-
-# Populate Chat
+# Early stages of sending message
 @dash.callback(
-    Output('table-body', 'children'),
-    Input('interval', 'n_intervals'),
-    Input('message-input', 'value'),
-    Input('userid-text', 'value'),
-    State('table-body', 'children'),
-    State('chat-id', 'value'),
-    prevent_initial_call=True
+  Output('messages', 'children'),
+  Input('send_message_btn', 'n_clicks'),
+  Input('send_message_field', 'n_submit'),
+  State('send_message_field', 'value'),
+  prevent_initial_call=True
 )
-def load(_, n_clicks, userid, children, chatid):
-    response = None
-    blocks = None
-    try:
-        response = util.sendRequest("get_room_messages", userid, chatid)
-        blocks = util.sendRequest("get_blocked_users", userid)
-    except:
-        return html.Div('An error occurred while running the startup script')
+def send_message(messages, button, field_submit):
+    
+    return f'{messages}' + f'{field_submit}'
 
-    data = response['data']
-    blocked_users = np.squeeze(blocks['data'])
-
-    # Create new child structure with messages
-    # Terribly inefficent
-    newChildren = []
-    for index in range(0, len(data) - 1, 4):
-        user = data[index]
-        message = data[index + 1]
-        timestamp = data[index + 2]
-        username = data[index + 3]
-
-        classes = "message-container message-"
-        classes = classes + "right" if user == userid else classes + "left"
-        if user in blocked_users:
-            classes += " message-blocked"
-        newChildren.append(html.Tr(children=[
-            html.Td(children=[
-                html.P(username, className='message-element message-username'),
-                html.P(message, className='message-element message-content'),
-                html.P(getTimestamp(timestamp), className='message-element message-timestamp')
-            ], className=classes)
-        ], className="message-row"))
-        
-    return newChildren
-
-# Send a message
+# Chat button callback
 @dash.callback(
-    # Output('error-text', 'children'),
-    Output('message-input', 'value'),
-    Input('message-input', 'value'),
-    State('userid-text', 'value'),
-    State('chat-id', 'value'),
-    prevent_initial_call=True
+  Output('recipient', 'children'),
+  [Input('{}-chatbtn'.format(people), 'n_clicks') for people in friends_list],
+  prevent_initial_call=True
 )
-def send_message(message, userid, groupid):
-    if message == '':
-        # return no_update, no_update
-        return no_update
-    message = pipes.quote(message)
+def set_recipient(*args):
+    changed_id = [p['prop_id'] for p in dash.callback_context.triggered][0]
+    # Output chat (current session, session clicked)
+    print(changed_id)
+    return changed_id.split('-')[0]
 
-    # If server fails to send message, retry
-    response = None
-    retryLimit = 10
-    trys = 0
-
-    while trys < retryLimit:
-        try:
-            response = util.sendRequest("create_message", userid, groupid, message)
-        except:
-            trys += 1
-            continue
-        break
-    if response:
-        return ''
-    return '[System] Message failed to send'
-
-def getTimestamp(timestamp):
-    # Calculate elapsed time for message sent to now
-    format = "%Y-%m-%d %H:%M:%S"
-    now = datetime.now()
-    message_time = datetime.strptime(timestamp, format)
-
-    now_seconds = now.timestamp()
-    message_seconds = message_time.timestamp()
-
-    elapsed_time = int(now_seconds - message_seconds)
-    # Format return
-    time_units = {'second': 60, 'minute' : 60, 'hour': 24, 'day': 7, 'week': 52, 'year': 10}
-
-    for i, unit in enumerate(list(time_units.keys())):
-        # Length of unit and previous unit in seconds
-        unit_len = np.prod(list(time_units.values())[0: i + 1])
-        punit_len = unit_len / time_units[unit]
-
-        # Elapsed time in unit format
-        elapsed_unit = math.floor(elapsed_time / punit_len)
-
-        if elapsed_time < unit_len:
-            postfix = ' ago' if elapsed_unit == 1 else 's ago'
-            return f'{elapsed_unit} {unit}{postfix}'
-
-# Display users
 @dash.callback(
-    Output('chatroom-table', 'children'),
-    Input('userid-text', 'value'),
-    prevent_initial_call=True
+  Output('friends-list', 'children'),
+  Output('blocked-list', 'children'),
+  [Input('{}-blockbtn'.format(people), 'n_clicks') for people in (friends_list + blocked_list)],
+  prevent_initial_call=True
 )
-def load_chatgroups(userid):
-    response = None
-    try:
-        response = util.sendRequest("get_user_rooms", userid)
-    except:
-        return html.Div('An error occurred while running the startup script')
+def block_unblock(*args):
+  changed_id = [p['prop_id'] for p in dash.callback_context.triggered][0]
+  button_id = changed_id.split('.')[0]
+  name = button_id.split('-')[0]
+  # Getting ID
 
-    chatrooms = response['data']
-    children = []
-    # global_chats = []
-    for room in chatrooms:
-        # global_chats.append(room[-2], room[-1])
-        # children.append(html.Div(className='chatroom-element', 
-        # children=[
-        #     html.Div(className="chatroom-element-text", children=[
-        #         html.P(children=[
-        #             html.Span("Area", id='chatroom-area-text'),
-        #             html.Span(": " + room[-2])
-        #         ], className='chatroom-element-area chatroom-element-text'),
-        #         html.P(children=[
-        #             html.Span("Time", id='chatroom-time-text'),
-        #             html.Span(": " + room[-1])
-        #         ], className='chatroom-element-time chatroom-element-text')
-        #     ]),
-        #     html.Div(id='chatroom-button-div', children=[
-        #         html.Button("View", id=f'chatroom-${room[-2]}:${room[-1]}', className="chatroom-view-button")
-        #     ])
-            
-        # ]))
+  # Guard against empty input
+  if changed_id == 'x':
+    return no_update, no_update
 
-        children.append(html.Div(className='chatroom', 
-        children=[
-            html.P(room[-2]),
-            html.P(room[-1])
-        ]))
+  if name in friends_list:
+    friends_list.remove(name) # SQL move to block
+    blocked_list.append(name) 
+  else:
+    blocked_list.remove(name) # SQL move to friends
+    friends_list.append(name)
 
-    return children;
-
-# Join group
-# @dash.callback(
-#     Output('chat-id', 'value'),
-#     [Input(f'chatroom-${value[0]}:${value[1]}', 'value') for _, value in enumerate(global_chats)],
-#     prevent_initial_call=True
-# )
-# def join_new_room(id, *args):
-#     return id
-
-# Set page layout
-def layout():
-    return success
+  # create_people_div_list(getFriends()/getBlocked())
+  return create_people_div_list(friends_list), create_people_div_list(blocked_list)
